# Vulkan: Add Integer Dot Product mul_mat_vec shader for legacy quants