… ggml/src/ggml-vulkan/vulkan-shaders/mul_mat_vec_base.comp
… ggml/src/ggml-vulkan/vulkan-shaders/mul_mat_vecq.comp
… ggml/src/ggml-vulkan/vulkan-shaders/mul_mmq.comp
… ggml/src/ggml-vulkan/vulkan-shaders/mul_mmq_funcs.comp
… ggml/src/ggml-vulkan/vulkan-shaders/quantize_q8_1.comp
… ggml/src/ggml-vulkan/vulkan-shaders/types.comp
… ggml/src/ggml-vulkan/vulkan-shaders/vulkan-shaders-gen.cpp