@0cc4m 2025-07-28
Yeah, I also see that. We might have to pick a threshold from which using this quantize + integer dot shader path is worth it. Even without further tuning, there are definitely cases where it helps, for example batch 4 and 8 on RX 6800 XT:

Master:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    512 |    1 |   1024 |    0.372 |  1378.10 |    6.499 |    78.78 |    6.871 |   149.04 |
|   512 |    512 |    2 |   2048 |    0.734 |  1394.93 |   11.341 |    90.29 |   12.075 |   169.60 |
|   512 |    512 |    4 |   4096 |    1.551 |  1320.62 |   18.337 |   111.69 |   19.887 |   205.96 |
|   512 |    512 |    8 |   8192 |    3.499 |  1170.69 |   34.641 |   118.24 |   38.139 |   214.79 |
|   512 |    512 |   16 |  16384 |    8.295 |   987.59 |   59.502 |   137.68 |   67.797 |   241.66 |
|   512 |    512 |   32 |  32768 |   21.548 |   760.35 |   85.820 |   190.91 |  107.368 |   305.19 |

PR:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    512 |    1 |   1024 |    0.372 |  1376.71 |    6.980 |    73.35 |    7.352 |   139.28 |
|   512 |    512 |    2 |   2048 |    0.721 |  1420.49 |   11.889 |    86.13 |   12.610 |   162.42 |
|   512 |    512 |    4 |   4096 |    1.562 |  1311.47 |   17.186 |   119.17 |   18.747 |   218.49 |
|   512 |    512 |    8 |   8192 |    3.482 |  1176.48 |   29.917 |   136.91 |   33.398 |   245.28 |
|   512 |    512 |   16 |  16384 |    8.253 |   992.55 |   59.530 |   137.61 |   67.783 |   241.71 |
|   512 |    512 |   32 |  32768 |   21.490 |   762.41 |   85.655 |   191.28 |  107.145 |   305.83 |