@0cc4m 2025-08-03
Here are some new results, performance is looking better now, even for small models. Selecting when to enable this and when not to is still tricky, though.

@jeffbolznv Can you retest on your worst-case 5090? On my 3090 it looks like enabling this path on Nvidia may be worth it on Q4_1 and Q5_1, since they perform best due to 16B alignment. If you see further optimization opportunities, let me know.

<details>
<summary>Nvidia RTX 3090 (without coopmat1/2)</summary>

ggml_vulkan: 0 = NVIDIA GeForce RTX 3090 (NVIDIA) | uma: 0 | fp16: 1 | bf16: 1 | warp size: 32 | shared memory: 49152 | int dot: 1 | matrix cores: none

| model                          |       size |     params | backend    | ngl |            test |         t/s (Master) |             t/s (PR) |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: | -------------------: |
| llama 1B Q4_0                  | 606.53 MiB |     1.10 B | Vulkan     |  99 |           pp512 |     9194.72 ± 323.10 |     8926.13 ± 203.39 |
| llama 1B Q4_0                  | 606.53 MiB |     1.10 B | Vulkan     |  99 |           tg128 |       324.21 ± 56.21 |       311.07 ± 51.04 |
| llama 1B Q4_1                  | 668.18 MiB |     1.10 B | Vulkan     |  99 |           pp512 |     9189.23 ± 148.56 |     9296.94 ± 194.27 |
| llama 1B Q4_1                  | 668.18 MiB |     1.10 B | Vulkan     |  99 |           tg128 |       336.64 ± 10.50 |        327.20 ± 0.56 |
| llama 1B Q5_0                  | 729.84 MiB |     1.10 B | Vulkan     |  99 |           pp512 |      8678.07 ± 32.60 |      9060.36 ± 21.48 |
| llama 1B Q5_0                  | 729.84 MiB |     1.10 B | Vulkan     |  99 |           tg128 |        304.93 ± 5.38 |        310.19 ± 4.96 |
| llama 1B Q5_1                  | 791.50 MiB |     1.10 B | Vulkan     |  99 |           pp512 |     8807.90 ± 204.32 |      9108.72 ± 30.17 |
| llama 1B Q5_1                  | 791.50 MiB |     1.10 B | Vulkan     |  99 |           tg128 |        303.30 ± 3.87 |        292.32 ± 0.86 |
| llama 1B Q8_0                  |   1.09 GiB |     1.10 B | Vulkan     |  99 |           pp512 |      9058.35 ± 32.73 |      9101.32 ± 23.69 |
| llama 1B Q8_0                  |   1.09 GiB |     1.10 B | Vulkan     |  99 |           tg128 |        288.87 ± 2.46 |        267.09 ± 2.26 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     |  99 |           pp512 |      1912.84 ± 15.65 |       1924.58 ± 9.18 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     |  99 |           tg128 |        107.09 ± 0.18 |        107.85 ± 0.75 |
| llama 8B Q4_1                  |   4.77 GiB |     8.03 B | Vulkan     |  99 |           pp512 |      1856.80 ± 10.88 |       1898.31 ± 9.23 |
| llama 8B Q4_1                  |   4.77 GiB |     8.03 B | Vulkan     |  99 |           tg128 |        101.17 ± 0.30 |        108.18 ± 0.15 |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | Vulkan     |  99 |           pp512 |      1884.42 ± 11.34 |       1898.32 ± 8.53 |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | Vulkan     |  99 |           tg128 |         75.10 ± 0.15 |         74.57 ± 0.11 |

```
Master:
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              20448 runs -    49.75 us/run - 117.44 MFLOP/run -   2.36 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              18744 runs -    54.75 us/run - 117.44 MFLOP/run -   2.15 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              14484 runs -    71.58 us/run - 117.44 MFLOP/run -   1.64 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              14484 runs -    72.32 us/run - 117.44 MFLOP/run -   1.62 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              12780 runs -    82.89 us/run - 117.44 MFLOP/run -   1.42 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              16188 runs -    62.95 us/run - 234.88 MFLOP/run -   3.73 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              12780 runs -    78.99 us/run - 234.88 MFLOP/run -   2.97 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              12354 runs -    83.76 us/run - 234.88 MFLOP/run -   2.80 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              10650 runs -    96.60 us/run - 234.88 MFLOP/run -   2.43 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              10224 runs -    98.03 us/run - 234.88 MFLOP/run -   2.40 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              12780 runs -    78.66 us/run - 352.32 MFLOP/run -   4.48 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              10224 runs -    98.22 us/run - 352.32 MFLOP/run -   3.59 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              10508 runs -    96.41 us/run - 352.32 MFLOP/run -   3.65 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               8804 runs -   114.84 us/run - 352.32 MFLOP/run -   3.07 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               7952 runs -   128.65 us/run - 352.32 MFLOP/run -   2.74 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              10650 runs -    94.40 us/run - 469.76 MFLOP/run -   4.98 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               8520 runs -   117.56 us/run - 469.76 MFLOP/run -   4.00 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               8946 runs -   113.66 us/run - 469.76 MFLOP/run -   4.13 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               7242 runs -   138.49 us/run - 469.76 MFLOP/run -   3.39 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               6390 runs -   158.86 us/run - 469.76 MFLOP/run -   2.96 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               8892 runs -   113.12 us/run - 587.20 MFLOP/run -   5.19 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               7182 runs -   141.11 us/run - 587.20 MFLOP/run -   4.16 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               7866 runs -   129.28 us/run - 587.20 MFLOP/run -   4.54 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               6327 runs -   158.58 us/run - 587.20 MFLOP/run -   3.70 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               5301 runs -   189.32 us/run - 587.20 MFLOP/run -   3.10 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               5992 runs -   166.94 us/run - 939.52 MFLOP/run -   5.63 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               4708 runs -   216.16 us/run - 939.52 MFLOP/run -   4.35 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               5564 runs -   181.81 us/run - 939.52 MFLOP/run -   5.17 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               4387 runs -   232.83 us/run - 939.52 MFLOP/run -   4.04 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               3638 runs -   277.63 us/run - 939.52 MFLOP/run -   3.38 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      452 runs -  2216.07 us/run -  60.13 GFLOP/run -  27.13 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      436 runs -  2294.29 us/run -  60.13 GFLOP/run -  26.21 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      430 runs -  2333.07 us/run -  60.13 GFLOP/run -  25.77 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      426 runs -  2354.50 us/run -  60.13 GFLOP/run -  25.54 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      450 runs -  2224.99 us/run -  60.13 GFLOP/run -  27.02 TFLOPS

PR:
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              23004 runs -    44.03 us/run - 117.44 MFLOP/run -   2.67 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              21300 runs -    47.81 us/run - 117.44 MFLOP/run -   2.46 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              18744 runs -    53.90 us/run - 117.44 MFLOP/run -   2.18 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              17892 runs -    56.25 us/run - 117.44 MFLOP/run -   2.09 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              12780 runs -    79.10 us/run - 117.44 MFLOP/run -   1.48 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              21300 runs -    47.81 us/run - 234.88 MFLOP/run -   4.91 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              20022 runs -    50.45 us/run - 234.88 MFLOP/run -   4.66 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              17892 runs -    56.73 us/run - 234.88 MFLOP/run -   4.14 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              17040 runs -    59.36 us/run - 234.88 MFLOP/run -   3.96 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              12354 runs -    81.78 us/run - 234.88 MFLOP/run -   2.87 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              19312 runs -    52.29 us/run - 352.32 MFLOP/run -   6.74 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              17892 runs -    56.05 us/run - 352.32 MFLOP/run -   6.29 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              16188 runs -    62.64 us/run - 352.32 MFLOP/run -   5.62 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              15620 runs -    64.26 us/run - 352.32 MFLOP/run -   5.48 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              11928 runs -    85.25 us/run - 352.32 MFLOP/run -   4.13 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              16827 runs -    59.72 us/run - 469.76 MFLOP/run -   7.87 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              16401 runs -    61.13 us/run - 469.76 MFLOP/run -   7.68 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              14271 runs -    70.17 us/run - 469.76 MFLOP/run -   6.69 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              14484 runs -    69.48 us/run - 469.76 MFLOP/run -   6.76 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=4,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              11076 runs -    91.45 us/run - 469.76 MFLOP/run -   5.14 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              14877 runs -    67.93 us/run - 587.20 MFLOP/run -   8.64 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              14193 runs -    71.19 us/run - 587.20 MFLOP/run -   8.25 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              12825 runs -    78.98 us/run - 587.20 MFLOP/run -   7.43 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              12996 runs -    77.27 us/run - 587.20 MFLOP/run -   7.60 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=5,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               9747 runs -   104.15 us/run - 587.20 MFLOP/run -   5.64 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              10486 runs -    95.93 us/run - 939.52 MFLOP/run -   9.79 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              10058 runs -    99.53 us/run - 939.52 MFLOP/run -   9.44 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               9309 runs -   107.87 us/run - 939.52 MFLOP/run -   8.71 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               9416 runs -   106.53 us/run - 939.52 MFLOP/run -   8.82 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=8,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):               6741 runs -   150.02 us/run - 939.52 MFLOP/run -   6.26 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      456 runs -  2200.07 us/run -  60.13 GFLOP/run -  27.33 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      450 runs -  2224.65 us/run -  60.13 GFLOP/run -  27.03 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      428 runs -  2342.33 us/run -  60.13 GFLOP/run -  25.67 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      438 runs -  2284.31 us/run -  60.13 GFLOP/run -  26.32 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):                      456 runs -  2196.48 us/run -  60.13 GFLOP/run -  27.38 TFLOPS
```

</details>

<details>
<summary>AMD Radeon RX 6800 XT</summary>

ggml_vulkan: 0 = AMD Radeon RX 6800 XT (RADV NAVI21) (radv) | uma: 0 | fp16: 1 | bf16: 0 | warp size: 32 | shared memory: 65536 | int dot: 1 | matrix cores: none
| model                          |       size |     params | backend    | ngl |            test |         t/s (Master) |             t/s (PR) |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: | -------------------: |
| llama 1B Q4_0                  | 606.53 MiB |     1.10 B | Vulkan     |  99 |           pp512 |      8290.44 ± 68.15 |     9310.68 ± 184.78 |
| llama 1B Q4_0                  | 606.53 MiB |     1.10 B | Vulkan     |  99 |           tg128 |        361.90 ± 0.40 |        346.92 ± 2.04 |
| llama 1B Q4_1                  | 668.18 MiB |     1.10 B | Vulkan     |  99 |           pp512 |      7996.84 ± 80.20 |     9143.00 ± 181.95 |
| llama 1B Q4_1                  | 668.18 MiB |     1.10 B | Vulkan     |  99 |           tg128 |        332.41 ± 0.99 |        333.82 ± 0.45 |
| llama 1B Q5_0                  | 729.84 MiB |     1.10 B | Vulkan     |  99 |           pp512 |      7788.58 ± 40.65 |     8975.54 ± 158.83 |
| llama 1B Q5_0                  | 729.84 MiB |     1.10 B | Vulkan     |  99 |           tg128 |        313.75 ± 0.63 |        317.04 ± 0.36 |
| llama 1B Q5_1                  | 791.50 MiB |     1.10 B | Vulkan     |  99 |           pp512 |      7645.57 ± 60.32 |     8879.91 ± 246.04 |
| llama 1B Q5_1                  | 791.50 MiB |     1.10 B | Vulkan     |  99 |           tg128 |        275.61 ± 0.57 |        299.83 ± 4.70 |
| llama 1B Q8_0                  |   1.09 GiB |     1.10 B | Vulkan     |  99 |           pp512 |      7207.73 ± 13.38 |     8179.54 ± 114.85 |
| llama 1B Q8_0                  |   1.09 GiB |     1.10 B | Vulkan     |  99 |           tg128 |        265.07 ± 0.14 |        249.98 ± 0.02 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     |  99 |           pp512 |       1460.24 ± 0.64 |       1651.88 ± 2.28 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     |  99 |           tg128 |         85.71 ± 0.05 |         86.02 ± 0.01 |
| llama 8B Q4_1                  |   4.77 GiB |     8.03 B | Vulkan     |  99 |           pp512 |       1421.29 ± 2.25 |       1602.68 ± 2.05 |
| llama 8B Q4_1                  |   4.77 GiB |     8.03 B | Vulkan     |  99 |           tg128 |         77.33 ± 0.15 |         79.11 ± 0.02 |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | Vulkan     |  99 |           pp512 |       1247.77 ± 0.86 |       1391.76 ± 1.91 |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | Vulkan     |  99 |           tg128 |         54.01 ± 0.06 |         53.82 ± 0.01 |


```
Master:
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              23004 runs -    45.06 us/run - 117.44 MFLOP/run -   2.61 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              18744 runs -    53.43 us/run - 117.44 MFLOP/run -   2.20 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              15336 runs -    67.74 us/run - 117.44 MFLOP/run -   1.73 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              14484 runs -    71.62 us/run - 117.44 MFLOP/run -   1.64 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              17892 runs -    56.38 us/run - 117.44 MFLOP/run -   2.08 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              17040 runs -    59.68 us/run - 234.88 MFLOP/run -   3.94 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              13206 runs -    75.97 us/run - 234.88 MFLOP/run -   3.09 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              12354 runs -    83.24 us/run - 234.88 MFLOP/run -   2.82 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              10650 runs -    94.06 us/run - 234.88 MFLOP/run -   2.50 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              11928 runs -    86.19 us/run - 234.88 MFLOP/run -   2.73 TFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              13632 runs -    73.55 us/run - 352.32 MFLOP/run -   4.79 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=3,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3],v=0):              10224 runs -    99.13 us/run - 352.32 MFLOP/run -   3.55 TFLOPS
…